Move tests, Add exception docs, rename Ticks to EpochTicks
